% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{.cancel_slurm_job}
\alias{.cancel_slurm_job}
\title{Cancel a SLURM job}
\usage{
.cancel_slurm_job(job_id)
}
\arguments{
\item{job_id}{Character. The SLURM job ID}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Cancels a SLURM job with the given job ID.
}
\keyword{internal}
