% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{.check_slurm_job_done}
\alias{.check_slurm_job_done}
\title{Check if a SLURM job is done}
\usage{
.check_slurm_job_done(job_id)
}
\arguments{
\item{job_id}{Character. The SLURM job ID}
}
\value{
Logical. TRUE if the job is done, FALSE otherwise.
}
\description{
Checks if a SLURM job with the given job ID is done.
}
\keyword{internal}
