% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.find_code_folder_name}
\alias{.find_code_folder_name}
\title{Find the name of the code folder in an analysis}
\usage{
.find_code_folder_name(analysis_path)
}
\arguments{
\item{analysis_path}{Character. Path to the analysis folder to inspect}
}
\description{
Find which name is used for the analysis' code folder.
This expects and validates that it is one of the known options.
}
\keyword{internal}
