% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.infer_analysis_from_path}
\alias{.infer_analysis_from_path}
\title{Infer an analysis from a file or directory path}
\usage{
.infer_analysis_from_path(start_path)
}
\description{
Infer an analysis from a file or directory path
}
\keyword{internal}
