% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{.validate_input_fields}
\alias{.validate_input_fields}
\title{Validate input fields}
\usage{
.validate_input_fields(input_def, input_id)
}
\arguments{
\item{input_def}{List. Input definition}

\item{input_id}{Character. Input ID for error messages}
}
\description{
Helper function to validate fingerprint fields in an input definition.
}
\keyword{internal}
