% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{pm_create_project}
\alias{pm_create_project}
\title{Create a new project from template in the given folder}
\usage{
pm_create_project(path)
}
\arguments{
\item{path}{Path to the folder in which to create the project}
}
\value{
\code{PMProject} object representing the newly created project
}
\description{
Creates a new clean project in the given folder and returns a
PMProject object.
Also works on an empty folder, but fails if the folder contains
irrelevant content.
}
\note{
If the folder already exists and has contents, does not override it,
but instead returns the existing project from that folder.
}
\examples{
empty_folder <- withr::local_tempdir()
pm <- pm_create_project(empty_folder)
pm

}
