% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pm_infer_analysis}
\alias{pm_infer_analysis}
\title{Infer an analysis object based on the calling script}
\usage{
pm_infer_analysis()
}
\value{
\code{PMAnalysis} object representing the inferred analysis
}
\description{
Find the relevant analysis object based on the call stack of the script
that invoked this function. Supported when called from a script in an
analysis folder inside the "analyses" folder, or from a script in the code
folder of an analysis (including nested subfolders). Falls back to the
\code{Rscript} entry file from command-line arguments, then to the current
working directory when the caller file cannot be determined otherwise.
}
\examples{
empty_folder <- withr::local_tempdir()
pm <- pm_create_project(empty_folder)
pm$create_analysis("my_analysis")

# Infer from analysis folder
withr::with_dir(file.path(empty_folder, "analyses", "my_analysis"), {
  analysis1 <- pm_infer_analysis()
  analysis1
})

# Infer from code folder
withr::with_dir(file.path(empty_folder, "analyses", "my_analysis", "code"), {
  analysis2 <- pm_infer_analysis()
  analysis2
})

}
