% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{pm_project}
\alias{pm_project}
\title{Create a PMProject object for a given project folder}
\usage{
pm_project(path)
}
\arguments{
\item{path}{Path to the project}
}
\value{
\code{PMProject} object representing the given project folder
}
\description{
Create a PMProject object for a given project folder
}
\examples{
# Create a valid project to load
folder <- withr::local_tempdir()
invisible(pm_create_project(folder))

# Load the project
pm <- pm_project(folder)
pm

}
