% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PMData}
\alias{PMData}
\title{Data input class}
\description{
Represents a single input data file with its ID and path.
This is a minimal representation that will be extended later
with reading/writing capabilities.
}
\examples{
data <- PMData$new(id = "feature_table", path = "/path/to/file.biom")
data$id
data$path


## ------------------------------------------------
## Method `PMData$read`
## ------------------------------------------------

\dontrun{
data <- PMData$new(id = "feature_table", path = "/path/to/file.csv")
contents <- data$read()
}

## ------------------------------------------------
## Method `PMData$write`
## ------------------------------------------------

withr::with_tempdir({
  data <- PMData$new(id = "feature_table", path = "file.csv")
  df <- data.frame(x = 1:5, y = letters[1:5])
  data$write(df)

  # For RData files, can write multiple objects
  data_rdata <- PMData$new(id = "results", path = "results.RData")
  obj1 <- data.frame(x = 1:3)
  obj2 <- c("a", "b", "c")
  data_rdata$write(obj1, obj2, obj3 = 42)
})
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Character. The canonical identifier for this input (e.g., "feature_table")}

\item{\code{path}}{Character. The absolute path to the input file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PMData-new}{\code{PMData$new()}}
\item \href{#method-PMData-print}{\code{PMData$print()}}
\item \href{#method-PMData-read}{\code{PMData$read()}}
\item \href{#method-PMData-write}{\code{PMData$write()}}
\item \href{#method-PMData-clone}{\code{PMData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PMData-new}{}}}
\subsection{Method \code{new()}}{
Create a PMData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMData$new(id, path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Character. The canonical identifier for this input}

\item{\code{path}}{Character. The absolute path to the input file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMData-print"></a>}}
\if{latex}{\out{\hypertarget{method-PMData-print}{}}}
\subsection{Method \code{print()}}{
Print method for PMData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMData-read"></a>}}
\if{latex}{\out{\hypertarget{method-PMData-read}{}}}
\subsection{Method \code{read()}}{
Read the data file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMData$read(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to \code{\link{pm_read_file}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The contents of the file. For RData files, returns a new environment
containing all objects from the file.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
data <- PMData$new(id = "feature_table", path = "/path/to/file.csv")
contents <- data$read()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMData-write"></a>}}
\if{latex}{\out{\hypertarget{method-PMData-write}{}}}
\subsection{Method \code{write()}}{
Write data to the file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMData$write(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The object to write. For tabular formats (CSV, TSV, Parquet), must be
a data.frame. For RData files, can accept multiple objects via \code{...}.}

\item{\code{...}}{Additional arguments passed to \code{\link{pm_write_file}}.
For RData files, additional objects can be provided here.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the file path.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{withr::with_tempdir({
  data <- PMData$new(id = "feature_table", path = "file.csv")
  df <- data.frame(x = 1:5, y = letters[1:5])
  data$write(df)

  # For RData files, can write multiple objects
  data_rdata <- PMData$new(id = "results", path = "results.RData")
  obj1 <- data.frame(x = 1:3)
  obj2 <- c("a", "b", "c")
  data_rdata$write(obj1, obj2, obj3 = 42)
})
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PMData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
