% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pm_infer_analysis}
\alias{pm_infer_analysis}
\title{Infer an analysis object based on the current directory}
\usage{
pm_infer_analysis()
}
\value{
\code{PMAnalysis} object representing the inferred analysis
}
\description{
Find the relevant analysis object based on the current directory.
Currently supported being called from an analysis folder inside
"analyses" folder, or from the "code" folder of an analysis.
}
\examples{
empty_folder <- withr::local_tempdir()
pm <- pm_create_project(empty_folder)
pm$create_analysis("my_analysis")

# Infer from analysis folder
withr::with_dir(file.path(empty_folder, "analyses", "my_analysis"), {
  analysis1 <- pm_infer_analysis()
  analysis1
})

# Infer from code folder
withr::with_dir(file.path(empty_folder, "analyses", "my_analysis", "code"), {
  analysis2 <- pm_infer_analysis()
  analysis2
})

}
