% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PMAnalysis}
\alias{PMAnalysis}
\title{Analysis class to manage an analysis folder}
\description{
This object can be used to manage an analysis folder within a project.
Each analysis contains code, outputs, intermediate results, and logs.
}
\examples{
# Create a project and analysis
folder <- withr::local_tempdir()
pm <- pm_create_project(folder)

# Create a new analysis
analysis <- pm$create_analysis("data_preparation")
analysis

# Load an existing analysis from project
analysis <- pm$get_analysis("data_preparation")

# Load an existing analysis from path
analysis <- PMAnalysis$new(path = file.path(folder, "analyses", "data_preparation"))


## ------------------------------------------------
## Method `PMAnalysis$get_output_path`
## ------------------------------------------------

folder <- withr::local_tempdir()
pm <- pm_create_project(folder)
analysis <- pm$create_analysis("my_analysis")

# Get output path for a CSV file
output <- analysis$get_output_path("results.csv", type = "table")
output$id    # "results"
output$path  # full path to results.csv in outputs/

# Get intermediate path without extension (will add .parquet for table type)
intermediate <- analysis$get_output_path("temp_data", type = "table", intermediate = TRUE)
intermediate$id    # "temp_data"
intermediate$path  # full path to temp_data.parquet in intermediate/
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Full path to the analysis's folder}

\item{\code{name}}{Name of the analysis (folder name within analyses/)}

\item{\code{project_path}}{Full path to the project's folder (if created from project)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PMAnalysis-new}{\code{PMAnalysis$new()}}
\item \href{#method-PMAnalysis-validate}{\code{PMAnalysis$validate()}}
\item \href{#method-PMAnalysis-print}{\code{PMAnalysis$print()}}
\item \href{#method-PMAnalysis-get_output_path}{\code{PMAnalysis$get_output_path()}}
\item \href{#method-PMAnalysis-clone}{\code{PMAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-new}{}}}
\subsection{Method \code{new()}}{
Create a PMAnalysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$new(project = NULL, name = NULL, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{PMProject object. If provided, creates analysis within that project.}

\item{\code{name}}{Character. Name of the analysis (folder name). Required if project is provided.}

\item{\code{path}}{Character. Full path to the analysis folder. Required if project is not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the analysis folder.
Makes sure all expected files and folders exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-print"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-print}{}}}
\subsection{Method \code{print()}}{
Print method for PMAnalysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-get_output_path"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-get_output_path}{}}}
\subsection{Method \code{get_output_path()}}{
Get output path for a file, returning a PMData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$get_output_path(name, type = NULL, intermediate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. Name of the output file (with or without extension).}

\item{\code{type}}{Character. Optional type of output (table, object, image, figure, parquet, csv).
If provided and name has no extension, an appropriate extension will be added.
If provided and name has an extension, the extension will be validated against the type.}

\item{\code{intermediate}}{Logical. If TRUE, file goes in intermediate/ folder; if FALSE, in outputs/ folder.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{PMData} object with:
\itemize{
\item \code{id}: The file name without extension
\item \code{path}: The full absolute path to the output file
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{folder <- withr::local_tempdir()
pm <- pm_create_project(folder)
analysis <- pm$create_analysis("my_analysis")

# Get output path for a CSV file
output <- analysis$get_output_path("results.csv", type = "table")
output$id    # "results"
output$path  # full path to results.csv in outputs/

# Get intermediate path without extension (will add .parquet for table type)
intermediate <- analysis$get_output_path("temp_data", type = "table", intermediate = TRUE)
intermediate$id    # "temp_data"
intermediate$path  # full path to temp_data.parquet in intermediate/
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
