% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PMAnalysis}
\alias{PMAnalysis}
\title{Analysis class to manage an analysis folder}
\description{
This object can be used to manage an analysis folder within a project.
Each analysis contains code, outputs, intermediate results, and logs.
}
\examples{
# Create a project and analysis
folder <- withr::local_tempdir()
pm <- pm_create_project(folder)

# Create a new analysis
analysis <- pm$create_analysis("data_preparation")
analysis

# Load an existing analysis from project
analysis <- pm$get_analysis("data_preparation")

# Load an existing analysis from path
analysis <- PMAnalysis$new(path = file.path(folder, "analyses", "data_preparation"))

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Full path to the analysis's folder}

\item{\code{name}}{Name of the analysis (folder name within analyses/)}

\item{\code{project_path}}{Full path to the project's folder (if created from project)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PMAnalysis-new}{\code{PMAnalysis$new()}}
\item \href{#method-PMAnalysis-validate}{\code{PMAnalysis$validate()}}
\item \href{#method-PMAnalysis-print}{\code{PMAnalysis$print()}}
\item \href{#method-PMAnalysis-clone}{\code{PMAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-new}{}}}
\subsection{Method \code{new()}}{
Create a PMAnalysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$new(project = NULL, name = NULL, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{PMProject object. If provided, creates analysis within that project.}

\item{\code{name}}{Character. Name of the analysis (folder name). Required if project is provided.}

\item{\code{path}}{Character. Full path to the analysis folder. Required if project is not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the analysis folder.
Makes sure all expected files and folders exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-print"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-print}{}}}
\subsection{Method \code{print()}}{
Print method for PMAnalysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMAnalysis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PMAnalysis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
