% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{PMProject}
\alias{PMProject}
\title{Project class to manage a project folder}
\description{
This object can be used to manage a project folder.
Use it to create analyses, freeze input files, etc.
}
\examples{
# Create a valid project to load
folder <- withr::local_tempdir()
invisible(pm_create_project(folder))

# Load the project
pm <- PMProject$new(folder)
pm


## ------------------------------------------------
## Method `PMProject$parse_inputs`
## ------------------------------------------------

folder <- withr::local_tempdir()
pm <- pm_create_project(folder)

# After configuring inputs.yaml and inputs.local.yaml:
# data_list <- pm$parse_inputs()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Full path to the project's folder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PMProject-new}{\code{PMProject$new()}}
\item \href{#method-PMProject-validate}{\code{PMProject$validate()}}
\item \href{#method-PMProject-parse_inputs}{\code{PMProject$parse_inputs()}}
\item \href{#method-PMProject-clone}{\code{PMProject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMProject-new"></a>}}
\if{latex}{\out{\hypertarget{method-PMProject-new}{}}}
\subsection{Method \code{new()}}{
Create a PMProject object in an existing project's folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMProject$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the project}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMProject-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PMProject-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the project folder.
Makes sure all expected files and folder exist and are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMProject$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMProject-parse_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-PMProject-parse_inputs}{}}}
\subsection{Method \code{parse_inputs()}}{
Parse inputs.yaml and inputs.local.yaml into PMData objects.
Reads the inputs definition file and the local paths file
and combines them to create a list of PMData objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMProject$parse_inputs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of PMData objects, one for each input defined in inputs.yaml
that has a corresponding path in inputs.local.yaml
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{folder <- withr::local_tempdir()
pm <- pm_create_project(folder)

# After configuring inputs.yaml and inputs.local.yaml:
# data_list <- pm$parse_inputs()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMProject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PMProject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PMProject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
